v0.12 (#10)
* Bump version to 0.12

* Additionally derive lock control values from Label

Some devices of type "Z-Wave Door Lock" do not identify their control pairs by the expected ControlUses 18 and 19, so in addition we now look for the ControlPair labels "Lock" and "Unlock" to find the ControlValue for these lock functions. No breaking change, this should allow currently non-working locks to work.